requirements.txt
aiohttp==3.8.1
boto3==1.24.77
cassandra-driver==3.25.0
cassandra-driver[graph]==3.25.0
fastapi==0.85.0
fuzzywuzzy==0.18.0
geomet==0.2.1.post1
gremlinpython==3.4.6
lxml==4.9.1
lz4==4.0.2
paho-mqtt==1.6.1
pathlib==1.0.1
psutil==5.9.2
psycopg2-binary==2.9.3
pygit2==1.10.1
requests==2.28.1
urllib3==1.26.12
uvicorn[standard]==0.18.3
